Update fixing controller and fixing mapping path in layout

+++ src/main/java/com/wira/EmployeeManagement/repository/RoleRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
-    Optional<Role> findById(Integer roleId);
+    Optional<Role> findById(Long roleId);
 }
